Move most of InputAccessor into libutil
