+++ src/libfetchers/input-accessor.cc
@@ -3,95 +3,6 @@
 
 namespace nix {
 
-static std::atomic<size_t> nextNumber{0};
-
-InputAccessor::InputAccessor()
-    : number(++nextNumber)
-{
-}
-
-// FIXME: merge with archive.cc.
-void InputAccessor::dumpPath(
-    const CanonPath & path,
-    Sink & sink,
-    PathFilter & filter)
-{
-    auto dumpContents = [&](const CanonPath & path)
-    {
-        // FIXME: pipe
-        auto s = readFile(path);
-        sink << "contents" << s.size();
-        sink(s);
-        writePadding(s.size(), sink);
-    };
-
-    std::function<void(const CanonPath & path)> dump;
-
-    dump = [&](const CanonPath & path) {
-        checkInterrupt();
-
-        auto st = lstat(path);
-
-        sink << "(";
-
-        if (st.type == tRegular) {
-            sink << "type" << "regular";
-            if (st.isExecutable)
-                sink << "executable" << "";
-            dumpContents(path);
-        }
-
-        else if (st.type == tDirectory) {
-            sink << "type" << "directory";
-
-            /* If we're on a case-insensitive system like macOS, undo
-               the case hack applied by restorePath(). */
-            std::map<std::string, std::string> unhacked;
-            for (auto & i : readDirectory(path))
-                if (/* archiveSettings.useCaseHack */ false) { // FIXME
-                    std::string name(i.first);
-                    size_t pos = i.first.find(caseHackSuffix);
-                    if (pos != std::string::npos) {
-                        debug("removing case hack suffix from '%s'", path + i.first);
-                        name.erase(pos);
-                    }
-                    if (!unhacked.emplace(name, i.first).second)
-                        throw Error("file name collision in between '%s' and '%s'",
-                            (path + unhacked[name]),
-                            (path + i.first));
-                } else
-                    unhacked.emplace(i.first, i.first);
-
-            for (auto & i : unhacked)
-                if (filter((path + i.first).abs())) {
-                    sink << "entry" << "(" << "name" << i.first << "node";
-                    dump(path + i.second);
-                    sink << ")";
-                }
-        }
-
-        else if (st.type == tSymlink)
-            sink << "type" << "symlink" << "target" << readLink(path);
-
-        else throw Error("file '%s' has an unsupported type", path);
-
-        sink << ")";
-    };
-
-    sink << narVersionMagic1;
-    dump(path);
-}
-
-Hash InputAccessor::hashPath(
-    const CanonPath & path,
-    PathFilter & filter,
-    HashType ht)
-{
-    HashSink sink(ht);
-    dumpPath(path, sink, filter);
-    return sink.finish().first;
-}
-
 StorePath InputAccessor::fetchToStore(
     ref<Store> store,
     const CanonPath & path,
@@ -112,19 +23,6 @@ StorePath InputAccessor::fetchToStore(
         : store->addToStoreFromDump(*source, name, method, htSHA256, repair);
 
     return storePath;
-}
-
-std::optional<InputAccessor::Stat> InputAccessor::maybeLstat(const CanonPath & path)
-{
-    // FIXME: merge these into one operation.
-    if (!pathExists(path))
-        return {};
-    return lstat(path);
-}
-
-std::string InputAccessor::showPath(const CanonPath & path)
-{
-    return path.abs();
 }
 
 SourcePath InputAccessor::root()
